xcsoar: Extract `xmlToJson()` function
